Move glue code for embassy-net into hophop
